+++ src/app/mod.rs
@@ -4,4 +4,5 @@ pub(crate) mod init;
 pub(crate) mod list;
 pub(crate) mod r#move;
 pub(crate) mod remove;
+pub(crate) mod rename;
 pub(crate) mod shell;