+++ src/cli/rename.rs
@@ -7,27 +7,29 @@ use clap::{Args, Subcommand};
     subcommand_value_name = "ACTION"
 )]
 pub struct RenameCommand {
-    /// Old alias name
-    #[arg()]
-    pub old: String,
-    /// New alias name
-    #[arg()]
-    pub new: String,
-    /// Rename a group instead of an alias
+    /// What to rename
     #[command(subcommand)]
-    pub group_rename: Option<GroupRename>,
+    pub target: RenameTarget,
 }
 
 #[derive(Subcommand)]
-pub enum GroupRename {
-    /// Rename a group
+pub enum RenameTarget {
+    /// Rename an existing alias
+    #[command(visible_alias = "a")]
+    Alias(RenameArgs),
+
+    /// Rename an existing group
     #[command(visible_alias = "g")]
-    Group {
-        /// Old group name
-        #[arg()]
-        old: String,
-        /// New group name
-        #[arg()]
-        new: String,
-    },
+    Group(RenameArgs),
+}
+
+#[derive(Args)]
+pub struct RenameArgs {
+    /// Current name
+    #[arg()]
+    pub old_name: String,
+
+    /// New name
+    #[arg()]
+    pub new_name: String,
 }