+++ src/core/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod edit;
 pub(crate) mod list;
 pub(crate) mod r#move;
 pub(crate) mod remove;
+pub(crate) mod rename;
 pub(crate) mod sync;
 
 /// Represents possible failure cases in core operations.